+++ Lib/Terminal/Service/Directories/CreateDirectoryService.php
@@ -4,7 +4,7 @@ namespace Lib\Terminal\Service\Directories;
 
 use DomainException;
 use Lib\Terminal\Domain\Directory;
-use Lib\Terminal\Infrastructure\TerminalDirectoriesDbContext;
+use Lib\Terminal\Infrastructure\Contexts\TerminalDirectoriesDbContext;
 
 class CreateDirectoryService
 {
