+++ Lib/Terminal/Service/Server/ReadServerService.php
@@ -4,7 +4,7 @@ namespace Lib\Terminal\Service\Server;
 
 use DI\NotFoundException;
 use Lib\Terminal\Domain\Server;
-use Lib\Terminal\Infrastructure\TerminalServerDbContext;
+use Lib\Terminal\Infrastructure\Contexts\TerminalServerDbContext;
 
 class ReadServerService
 {
