+++ Lib/Terminal/Service/Server/UpdateServerService.php
@@ -3,7 +3,7 @@
 namespace Lib\Terminal\Service\Server;
 
 use Lib\Terminal\Domain\Server;
-use Lib\Terminal\Infrastructure\TerminalServerDbContext;
+use Lib\Terminal\Infrastructure\Contexts\TerminalServerDbContext;
 
 class UpdateServerService
 {
